requirements.txt
langchain==0.1.13
openai==1.14.3
python-dotenv==1.0.1
streamlit==1.24.1
streamlit_chat==0.0.2.2
